requirements.txt
aiosmtplib==3.0.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
bcrypt==4.0.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Deprecated==1.2.18
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.11
fastapi-cache2==0.2.2
fastapi-limiter==0.1.6
fastapi-mail==1.4.2
gevent==25.4.1
greenlet==3.2.0
gunicorn==23.0.0
h11==0.16.0
httptools==0.6.4
idna==3.10
inotify==0.2.10
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.0.0
Mako==1.3.10
MarkupSafe==3.0.2
mmh3==5.1.0
nose==1.3.7
packaging==24.2
passlib==1.7.4
pendulum==3.1.0
pillow==11.1.0
pottery==3.0.1
psycopg2==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==4.6.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.1
typing_extensions==4.12.2
tzdata==2025.2
uvicorn==0.34.0
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.0.5
websocket==0.2.1
websockets==15.0.1
wrapt==1.17.2
zope.event==5.0
zope.interface==7.2
